requirements.txt
matplotlib==3.5.1
numpy==1.22.2
pandas==1.4.1
Pillow==9.1.1
plotly==5.6.0
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.5.1
